Moved FirebaseTaskIdentity to exclusive module
