+++ src/auto_flutter_dig/task/identity/__init__.py
@@ -1,5 +1,5 @@
 from ...module.aflutter.identity import AflutterTaskIdentity
-from .firebase import FirebaseTaskIdentity
+from ...module.firebase.identity import FirebaseTaskIdentity
 from .flutter import FlutterTaskIdentity
 
 __all__ = ["AflutterTaskIdentity", "FirebaseTaskIdentity", "FlutterTaskIdentity"]
